Bundle libhif
There's major work going on in the "libhif" project to merge a lot of
functionality.  Things are going to be unstable for a while, and it's
not realistic to tie the release cycles of the consumers (rpm-ostree,
dnf, PackageKit) all together at once.

We'll iterate and break API for a while, then try to stabilize later.