tests: run parts of the vmcheck testsuite
Run the basic and layering-basic vmcheck tests on qcow2 Atomic Host
images. It's full of hacks for now; I'd like to confirm first that there
aren't any oversights that would prevent the pipeline from running it.

Once we get it going, we can work on cleaning up and upstreaming the
hacks to standard-test-roles, and run all the vmcheck tests.