tests: just directly rsync files
Using the synchronize module made rsync try to fetch the files from the
VM, even though we were part of the delegated block. Might be a subtle
bug in ansible. But really, there's no point in using the module here.
Both src and dest are local, so we can just call rsync ourselves.